ends_with_6
This program asks the user for an input, then ask another if the first input ends with the second input. This will run trough a for-loop then appends to a list then it will be reversed, to use the startswith() function. This acts as a endswith() function without using it.